rest cotrollers plus service complete, still implementing the restcontrollers

+++ src/main/java/com/csaba79coder/littersnap/model/report/service/ReportService.java
@@ -20,24 +20,58 @@ public class ReportService  {
 
     private final ReportRepository reportRepository;
 
-    public List<ReportModel> findAllReports() {
+
+    public List<ReportModel> getAllReports() {
         return reportRepository.findAll()
                 .stream()
                 .map(Mapper::mapReportEntityToModel)
                 .collect(Collectors.toList());
     }
 
-    public ReportModel findReportByID(UUID id) {
+    public ReportModel addNewReport(ReportModel reportModel) {
+        Report reportEntity = Mapper.mapReportModelToEntity(reportModel);
+        Report savedReportEntity = reportRepository.save(reportEntity);
+        return Mapper.mapReportEntityToModel(savedReportEntity);
+    }
+
+
+    public ReportModel getReportById(UUID id) {
         Optional<Report> optionalReport = reportRepository.findReportById(id);
-        return optionalReport.map(Mapper::mapReportEntityToModel).orElse(null);
+        if (optionalReport.isPresent()) {
+            Report report = optionalReport.get();
+            return Mapper.mapReportEntityToModel(report);
+        }
+        //TODO
+        // If the report with the specified ID does not exist, you can throw an exception or return null
+        return null;
     }
 
-    public ReportModel addNewReport(Report report) {
-        return Mapper.mapReportEntityToModel(reportRepository.save(report));
+    public ReportModel updateExistingReport(UUID id, ReportModel model) {
+        Optional<Report> optionalExistingReport = reportRepository.findById(id);
+        if (optionalExistingReport.isPresent()) {
+            Report existingReport = optionalExistingReport.get();
+
+            // Update the properties of the existing report with the values from the model
+            existingReport.setLitter(model.getLitter());
+
+            // Save the updated report in the repository
+            Report updatedReport = reportRepository.save(existingReport);
+
+            // Map the updated report entity to a ReportModel and return it
+            return Mapper.mapReportEntityToModel(updatedReport);
+        }
+        //TODO
+        // If the report with the specified ID does not exist, you can throw an exception or return null
+        return null;
     }
 
-    public void deleteReportById(UUID id) {
-        reportRepository.deleteById(id);
+    public void deleteExistingReport(UUID id) {
+        Optional<Report> optionalExistingReport = reportRepository.findById(id);
+        if (optionalExistingReport.isPresent()) {
+            reportRepository.deleteById(id);
+        } else {
+            //TODO
+            // If the report with the specified ID does not exist, you can throw an exception or return null
+        }
     }
-
 }
